# Implement jagged argsort and speed up argminmax