--- a/awkward/version.py
+++ b/awkward/version.py
@@ -4,7 +4,7 @@
 
 import re
 
-__version__ = "0.12.2"
+__version__ = "0.12.3"
 version = __version__
 version_info = tuple(re.split(r"[-\.]", __version__))