At first, I enabled the 'staircase' method for `argminmax` in 322ad114e359b50c3663f7ffa016d5991fb02156, but realized it was susceptible to precision issues, e.g. the following code would produce incorrect output (should be all 1s):
```python
import numpy as np, awkward
n = 10
a = awkward.JaggedArray.fromcounts(np.full(n, 2), np.tile(np.array([2**40, 2**40 + 2**-12]), n))
a.argmax().tolist()
```
It turns out that for large enough arrays, `a.localindex[a.max() == a]` is actually about as performant as the staircase method.

Then, using these argmax in a basic N^2 sort routine, one can get 'reasonable' performance for argsort, e.g.
```python
n = 100000
mu = 3
counts = np.random.poisson(mu, size=n)
content = np.random.normal(size=counts.sum())
a = awkward.JaggedArray.fromcounts(counts, content)
%timeit a.argsort()
%timeit a.copy(content=np.argsort(a.content + a.content.max()*a.parents) - a.starts[a.parents])
```
produces
```
313 ms ± 11.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
9.45 ms ± 306 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
so only 30x slower.  There is room for more micro-optimization in the implementation.

Fixes #81