@jpivarski 2019-07-15
> `a.localindex[a.max() == a]`

Isn't it possible for two elements to have the same maximum? What happens then?