@nsmith- 2019-07-15
In the case of argmax, take the first (as before) and for sort, it acts stably