@jpivarski 2019-07-15
It's a mask—doesn't it select both? If it's testing elements of the inner arrays, it can't be a flat boolean array—it must be a jagged boolean array. (Although you can force it to take at most one with `[:1]`...)

Also, if you've tested it and I just can't do indexing in my head, that's fair, too.

Oh, one last thing—what about `nan`? That's not equal to itself.