@nsmith- 2019-07-15
For argmax, yes I take at most one, which is what the lines in https://github.com/scikit-hep/awkward-array/blob/db7933c0dac8c413ce3a3292763d3f14429110c9/awkward/array/jagged.py#L1528-L1536 do.  The PR diff looks awful so you might not have seen that part.

For NaNs, already I observe that NaN does not propagate up from awkward reducers (bug?) so I decide to leave them alone for argminmax.  For sorting, I push them all to the end of the array (regardless of ascending/descending) with https://github.com/scikit-hep/awkward-array/blob/db7933c0dac8c413ce3a3292763d3f14429110c9/awkward/array/jagged.py#L1554