@jpivarski 2019-07-15
The policy that reducers take is to skip `nans` entirely, so they can never be the max or the min of an inner array. An array of all `nan` is as good as an empty array.

Note: the same is true of missing values due to MaskedArrays. A first order thing to think about is whether this breaks that behavior for reducers and argmax/argmin, which should have been recently tested in the homogeneity sweep, but not converted into automated tests yet. A second order thing to think about is what it would take to do the same for sort and argsort.

For sorting, users expect it to not change the length of the array, so they should be pushed to the end or something. (It could be the beginning or the end, but we like putting the most relevant values at the beginning, and `nans` are the least relevant.)

I'm on vacation and don't have easy access to look at the code right now. I could try it tomorrow morning (by running the tutorial I'm preparing for CoDaS-HEP before everybody wakes up).