@nsmith- 2019-07-15
Re: reducer policy - fine by me, but FYI
```python
>>> np.max(np.array([5., 2., np.nan, 4.]))
nan
>>> np.min(np.array([5., 2., np.nan, 4.]))
nan
```

Re:
> A first order thing to think about is whether this breaks that behavior for reducers and argmax/argmin, which should have been recently tested in the homogeneity sweep, but not converted into automated tests yet. A second order thing to think about is what it would take to do the same for sort and argsort.

So I would recommend adding unit tests as you check things.  Anyway, `argmax()` did not handle masked arrays correctly but now does:
```python
>>> import awkward
>>> a = awkward.fromiter([[4., 2., None], [3., 5.], [None]])
>>> a.argmin()
<JaggedArray [[1] [0] []] at 0x00010da95c18>
>>> a.argmax()
<JaggedArray [[0] [1] []] at 0x00010ce8c630>
```
however the new `argsort` doesn't handle it.  Should a masked element disappear under `argsort` like it does under `__getitem__` or should it go to the end of the array (after NaN, for example)?  For the first, the solution is easy, and for the second I think there is a solution but trying it uncovered another bug:
```python
>>> ~(a.min()==a).fillna(True)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/ncsmith/src/awkward-array/awkward/array/base.py", line 719, in fillna
    return self.copy(content=self._util_fillna(self._content, value))
  File "/Users/ncsmith/src/awkward-array/awkward/array/base.py", line 634, in _util_fillna
    return array.fillna(value)
  File "/Users/ncsmith/src/awkward-array/awkward/array/masked.py", line 889, in fillna
    out[self.mask < 0] = value
IndexError: boolean index did not match indexed array along dimension 0; dimension is 4 but corresponding boolean dimension is 6
```