@nsmith- 2019-07-16
Fixes #81