@jpivarski 2019-07-16
It sounds like the first order is not only okay, but fixed by you, so thanks! The second order (have sorting deal with `None`) doesn't need to be done yet.