@nsmith- 2019-07-16
I put something in to at least not give undefined behavior: masked items are dropped from the resulting index list.  This means that `a[a.argsort()]` would produce a sorted output with the masked entries dropped.