@jpivarski 2019-07-17
I brought this PR up to date and bumped the version to prepare for a deployment. I've tested the merged version with uproot and the tutorial I'm writing—everything's good—so if these tests pass, I'll merge it.